Use debian:jessie for smaller base image and forego instead of the init system, image size down to ~200MB from ~500MB

+++ assets/scripts/start.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
 mkdir -p /data/{transmission,watch,downloads}
-cp /opt/transmission.json /data/transmission/settings.json
-test ! -L /srv/http/downloads && ln -s /data/downloads /srv/http/downloads
-test -n "$EMAIL" && echo Environment="EMAIL=$EMAIL" >> /etc/systemd/system/transmission.service.d/custom.conf
+test ! -f /data/transmission/settings.json && cp /opt/transmission.json /data/transmission/settings.json
+test ! -L /var/www/html/downloads && ln -s /{data,var/www/html}/downloads
+
+mkdir -p /run/minidlna
 
 list() { l=`echo $@ | sed 's/,/" "/g'`; echo "(\"$l\")"; }
 eval FEEDS=`list $RSS_FEED`
@@ -13,8 +14,8 @@ rm -f /opt/tv.cron
 for i in "${!FEEDS[@]}"; do
     feed="${FEEDS[$i]}"
     opts="${OPTS[$i]}"
-    echo "30 4 * * * /opt/scripts/tv.sh -l $feed -o /data/watch $opts" >> /opt/tv.cron
+    test -n "$feed" && echo "30 4 * * * /opt/scripts/tv.sh -l $feed -o /data/watch $opts" >> /opt/tv.cron
 done
-crontab /opt/tv.cron
+test -f /opt/tv.cron && crontab /opt/tv.cron
 
-exec /usr/bin/init
+exec "$@"
